+++ src/lib.rs
@@ -21,6 +21,7 @@ pub mod indexer;
 pub mod kmd;
 
 pub mod atomic_transaction_composer;
+pub mod dryrun;
 
 pub mod error;
 pub use error::Error;